move query distribution folder to cluster
