+++ cluster/src/main/java/org/apache/iotdb/cluster/query/distribution/common/FillPolicy.java
@@ -1,0 +1,5 @@
+package org.apache.iotdb.cluster.query.distribution.common;
+
+public enum FillPolicy {
+    PREVIOUS,
+}
