+++ cluster/src/main/java/org/apache/iotdb/cluster/query/distribution/common/WithoutPolicy.java
@@ -1,4 +1,4 @@
-package org.apache.iotdb.db.query.distribution.common;
+package org.apache.iotdb.cluster.query.distribution.common;
 
 public enum WithoutPolicy {
     CONTAINS_NULL,
