+++ cluster/src/main/java/org/apache/iotdb/cluster/query/distribution/operator/WithoutOperator.java
@@ -1,7 +1,7 @@
-package org.apache.iotdb.db.query.distribution.operator;
+package org.apache.iotdb.cluster.query.distribution.operator;
 
-import org.apache.iotdb.db.query.distribution.common.Tablet;
-import org.apache.iotdb.db.query.distribution.common.WithoutPolicy;
+import org.apache.iotdb.cluster.query.distribution.common.Tablet;
+import org.apache.iotdb.cluster.query.distribution.common.WithoutPolicy;
 
 /**
  * WithoutOperator is used to discard specific result from upstream operators.
