--- server/src/main/java/org/apache/iotdb/db/query/distribution/common/FillPolicy.java
@@ -1,5 +1,0 @@
-package org.apache.iotdb.db.query.distribution.common;
-
-public enum FillPolicy {
-    PREVIOUS,
-}
